requirements.txt
aiogram==3.4.1
gspread==5.12.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
APScheduler==3.10.4
loguru==0.7.2
python-dotenv==1.0.0
aiosqlite==0.19.0
pytz==2024.1

